修改一些错误的描述，改了下剪枝逻辑。

+++ Abp_TicTacToe.h
@@ -3,7 +3,7 @@
 #include "TicTaeToeBoard.h"
 #include <stdio.h>
 
-double minimax(TicTacToeBoard *const, double, double, bool);
+int minimax(TicTacToeBoard *const, int, int, bool);
 int ai_move(TicTacToeBoard *const);
 int get_human_input(const TicTacToeBoard *const);
